processing/src/requirements.txt

beautifulsoup4==4.11.1
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
colorama==0.4.6
comm==0.1.2
cryptography==38.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
flit_core==3.6.0
python-dotenv
idna==3.4
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
lxml==4.9.1
matplotlib-inline==0.1.6
mistune==0.8.4
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pip==22.3.1
platformdirs==2.5.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==8.0.0
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
Send2Trash==1.8.0
setuptools==65.5.0
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spark-nlp==4.2.6
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
